requirements.txt
requests==2.25.1
discord.py==1.7.1
firebase-admin==4.5.3
discord-ext-alternatives==2021.4.13
DiscordUtils==1.2.9.1
asyncio==3.4.3
datetime==4.3
pychal==1.10.0
NHentai-API==0.0.15
jishaku==2.0.0
python-dotenv==0.17.0